Refactor main.py into modular route files
Split the 4800+ line main.py into 16 route modules and 3 utility modules:
- routes/: results, donors, members, activities, events, engagement,
  gallery, training_tips, credits, banners, homepage, uploads,
  meeting_minutes, recurrence, highlights, settings
- utils/: auth.py, s3.py, time.py

main.py is now ~130 lines: app init, middleware, lifespan, router includes.
All 136 endpoints tested and working. Updated CLAUDE.md with new structure.
